Enum extensions

+++ Items/Log/LogEntry.cs
@@ -8,7 +8,7 @@
         public string? userID { get; set; }
         public string? username { get; set; }
         public string? accessLocation { get; set; }
-        public int accessDirection { get; set; }
+        public UInt16 accessDirection { get; set; }
         public UInt16 verifyStatusCode { get; set; }
         public string? additionalInfo { get; set; }
         public string? logTime { get; set; }
